More specific instructions for contacting interrupt
Outside of the core Cloud Foundry core contributor community, people may not realize that "interrupt" means the people who should be contacted. This is an attempt to give more specific instructions, but there's probably still room for improvement.  Feel free to just use this as inspiration if you have a better idea of how to word it.